Feature/fix memory leak (#1)
* Init go module

* Fix bluestein cache memory leak through lfu strategy cache with 64 items:
According to mjibson#23 there is a memory leak bug in the caching of bluestein factors. This addresses this through an lfu cache with a capacity of 64 items. Based on the input length and thus the size of the computed factors this is sufficient.
The max size of the cache can be calculated through:
size_factor = (128*2) / 8 = 32
max_size = size_factor * input_len * 64

E.g.: A 1 sec wav file with sample rate of 44.1 kHz and an arbitrary sample size results in
32 * (1 * 44100) * 64 = 86.13 MByte